@@ -1,5 +1,4 @@
-import 'unplugin-icons/types/svelte';
-// See https://kit.svelte.dev/docs/types#app
+// See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 declare global {
 	namespace App {
